// frontend/src/model/Upgrade.ts
export type Upgrade = {
    username: string;
    click1upgrade: number;
    click2upgrade: number;
    click3upgrade: number;
    click4upgrade: number;
    click5upgrade: number;
    income1upgrade: number;
    income2upgrade: number;
    income3upgrade: number;
    income4upgrade: number;
    income5upgrade: number;
}